Update Java code generators
