+++ codegen/src/main/scala/akka/grpc/gen/javadsl/JavaClientCodeGenerator.scala
@@ -12,8 +12,7 @@ import templates.JavaClient.txt.Client
 trait JavaClientCodeGenerator extends JavaCodeGenerator {
   override def name = "akka-grpc-javadsl-client"
 
-  override def perServiceContent: Set[(Logger, Service) ⇒ CodeGeneratorResponse.File] = super.perServiceContent +
-    JavaCodeGenerator.generateServiceFile + generateStub
+  override def perServiceContent: Set[(Logger, Service) ⇒ CodeGeneratorResponse.File] = super.perServiceContent + generateStub
 
   def generateStub(logger: Logger, service: Service): CodeGeneratorResponse.File = {
     val b = CodeGeneratorResponse.File.newBuilder()
