+++ codegen/src/main/scala/akka/grpc/gen/javadsl/JavaServerCodeGenerator.scala
@@ -12,8 +12,7 @@ import templates.JavaServer.txt.Handler
 trait JavaServerCodeGenerator extends JavaCodeGenerator {
   override def name = "akka-grpc-javadsl-server"
 
-  override def perServiceContent: Set[(Logger, Service) ⇒ CodeGeneratorResponse.File] = super.perServiceContent +
-    JavaCodeGenerator.generateServiceFile + generateHandlerFactory
+  override def perServiceContent: Set[(Logger, Service) ⇒ CodeGeneratorResponse.File] = super.perServiceContent + generateHandlerFactory
 
   def generateHandlerFactory(logger: Logger, service: Service): CodeGeneratorResponse.File = {
     val b = CodeGeneratorResponse.File.newBuilder()
